Perform cold block heapification of instances of value types
If a candidate for stack allocation is detected to escape only in cold
blocks, Escape Analysis might still allow it to be stack allocated with
compensatory code in cold blocks that will copy it to an object on the
heap, known as cold block heapification.  This change adds cold block
heapification support for instances of value types.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>